Rework normalize to handle implicit accepts by a recursive wildcard higher up in the tree. So for instance:
+**
+/a/b

Since ** and b are not on the same level, this was not caught until now. "+/a/b" is now eliminated.

As part of this change, we now also prune paths that end without acceptance. This
would happen when an accepting node was eliminated - the elimination was not
propagated backwards until now.